requirements.txt
asgiref==3.4.1
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-colorfield==0.4.3
django-cors-headers==3.10.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
drf-extra-fields==3.1.1
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.3
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
